+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Change Log
 =============
 
-[4.3.8RC] - 2022-03-06
+[4.4.0] - 2022-03-10
 --------------------
 - [CHANGED] introduce a single-node option for running a case study as a single node (no network constraints).
 - [CHANGED] new option value 2 for IndBinGenInvest, IndBinGenRetirement, IndBinNetInvest for ignoring the investment/retirement decisions.